Merged PR 3749: - FEAT: Added client-hints functionality to the module.
- FEAT: Added client-hints functionality to the module.
This adds a `set_ch_headers` function which requests that the browser sends client-hint headers. Unit tests and an example have also been added.
As client-hints are only available in the Enterprise data file, the unit tests will be skipped if the Lite data file is used. The data file to use is now an optional configuration setting.
- TEST: Tests which rely on the "test" configuration being used to build now only run if that is true, otherwise they are skipped.